requirements.txt
requests
shodan
mmh3
dnspython
pyOpenSSL
wafw00f
